+++ Common/src/main/java/com/nyfaria/spookybats/entity/UndeadBat.java
@@ -13,6 +13,8 @@ import net.minecraft.world.entity.projectile.LargeFireball;
 import net.minecraft.world.level.Level;
 
 public class UndeadBat extends FireProofBat implements RangedAttackMob {
+	private int shootCooldown = 1;
+
 	public UndeadBat(EntityType<? extends SpookyBat> entityType, Level level) {
 		super(entityType, level);
 	}
@@ -27,19 +29,25 @@ public class UndeadBat extends FireProofBat implements RangedAttackMob {
 
 	@Override
 	public void performRangedAttack(LivingEntity target, float velocity) {
-		double $$5 = target.getX() - (this.getX() + this.getViewVector(1).x * 4.0);
-		double $$6 = target.getY(0.5) - (0.5 + this.getY(0.5));
-		double $$7 = target.getZ() - (this.getZ() + this.getViewVector(1).z * 4.0);
+		if (shootCooldown % 2 == 0) {
+			double $$5 = target.getX() - (this.getX() + this.getViewVector(1).x * 4.0);
+			double $$6 = target.getY(0.5) - (0.5 + this.getY(0.5));
+			double $$7 = target.getZ() - (this.getZ() + this.getViewVector(1).z * 4.0);
 
-		LargeFireball fireball = new LargeFireball(this.level(), this, $$5, $$6, $$7, 0);
+			LargeFireball fireball = new LargeFireball(this.level(), this, $$5, $$6, $$7, 0);
 
-		double d0 = target.getX() - this.getX();
-		double d1 = target.getY(0.3333333333333333D) - fireball.getY();
-		double d2 = target.getZ() - this.getZ();
-		double d3 = Math.sqrt(d0 * d0 + d2 * d2);
+			double d0 = target.getX() - this.getX();
+			double d1 = target.getY(0.3333333333333333D) - fireball.getY();
+			double d2 = target.getZ() - this.getZ();
+			double d3 = Math.sqrt(d0 * d0 + d2 * d2);
 
-		fireball.shoot(d0, d1 + d3 * (double) 0.2F, d2, 1.6F, (float) (14 - this.level().getDifficulty().getId() * 4));
-		this.playSound(SoundEvents.BLAZE_SHOOT, 1.0F, 1.0F / (this.getRandom().nextFloat() * 0.4F + 0.8F));
-		this.level().addFreshEntity(fireball);
+			fireball.shoot(d0, d1 + d3 * (double) 0.2F, d2, 1.6F, (float) (14 - this.level().getDifficulty().getId() * 4));
+			this.playSound(SoundEvents.BLAZE_SHOOT, 1.0F, 1.0F / (this.getRandom().nextFloat() * 0.4F + 0.8F));
+			this.level().addFreshEntity(fireball);
+
+			this.shootCooldown = 1;
+		} else {
+			this.shootCooldown++;
+		}
 	}
 }