+++ Fabric/src/main/java/com/nyfaria/spookybats/SpookyBatsClient.java
@@ -33,7 +33,7 @@ public class SpookyBatsClient implements ClientModInitializer {
         EntityRendererRegistry.register(EntityInit.EVIL_BAT.get(), context -> new SpookyBatRenderer<>(context, new EvilBatModel<>(context.bakeLayer(EvilBatModel.LAYER_LOCATION))));
         EntityRendererRegistry.register(EntityInit.VOID_BAT.get(), context->new VoidBatRenderer(context, new HatBatModel<>(context.bakeLayer(HatBatModel.LAYER_LOCATION))));
         EntityRendererRegistry.register(EntityInit.SLIME_BAT.get(), context -> new SlimeBatRenderer(context, new SlimeBatModel<>(context.bakeLayer(SlimeBatModel.LAYER_LOCATION))));
-
+        EntityRendererRegistry.register(EntityInit.EXPERIENCE_ORB_BAT.get(), context -> new EmissiveBatRenderer(context, new ExperienceOrbBatModel<>(context.bakeLayer(ExperienceOrbBatModel.LAYER_LOCATION)), "experience_orb_bat"));
 
         EntityRendererRegistry.register(EntityInit.JACK_O_LANTERN_PROJECTILE.get(), context -> new ThrownItemRenderer<>(context, 1.0f,true));
 
@@ -47,6 +47,8 @@ public class SpookyBatsClient implements ClientModInitializer {
         EntityModelLayerRegistry.registerModelLayer(GhostBatModel.LAYER_LOCATION, GhostBatModel::createBodyLayer);
         EntityModelLayerRegistry.registerModelLayer(EvilBatModel.LAYER_LOCATION, EvilBatModel::createBodyLayer);
         EntityModelLayerRegistry.registerModelLayer(SlimeBatModel.LAYER_LOCATION, SlimeBatModel::createBodyLayer);
+        EntityModelLayerRegistry.registerModelLayer(ExperienceOrbBatModel.LAYER_LOCATION, ExperienceOrbBatModel::createBodyLayer);
+
 
         LivingEntityFeatureRendererRegistrationCallback.EVENT.register((entityType, entityRenderer, registrationHelper, context) -> {
             if (entityType == EntityType.PLAYER) {