+++ Common/src/main/java/com/nyfaria/spookybats/client/model/ShulkerBatModel.java
@@ -1,0 +1,47 @@
+package com.nyfaria.spookybats.client.model;
+
+import com.nyfaria.spookybats.entity.api.SpookyBat;
+import net.minecraft.client.model.geom.ModelLayerLocation;
+import net.minecraft.client.model.geom.ModelPart;
+import net.minecraft.client.model.geom.PartPose;
+import net.minecraft.client.model.geom.builders.*;
+import net.minecraft.resources.ResourceLocation;
+
+public class ShulkerBatModel<T extends SpookyBat> extends SpookyBatModel<T> {
+	public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(new ResourceLocation("modid", "shulker_bat"), "main");
+
+	public ShulkerBatModel(ModelPart pRoot) {
+		super(pRoot);
+	}
+
+	public static LayerDefinition createBodyLayer() {
+		MeshDefinition meshdefinition = new MeshDefinition();
+		PartDefinition partdefinition = meshdefinition.getRoot();
+
+		PartDefinition head = partdefinition.addOrReplaceChild("head", CubeListBuilder.create(), PartPose.offset(-0.25F, 9.5F, 2.5F));
+
+		PartDefinition right_ear = head.addOrReplaceChild("right_ear", CubeListBuilder.create(), PartPose.offset(0.25F, 14.5F, -2.5F));
+
+		PartDefinition left_ear = head.addOrReplaceChild("left_ear", CubeListBuilder.create(), PartPose.offset(-0.75F, -7.5F, -3.5F));
+
+		PartDefinition shulker = head.addOrReplaceChild("shulker", CubeListBuilder.create(), PartPose.offset(0.25F, 9.5F, 0.5F));
+
+		PartDefinition lid = shulker.addOrReplaceChild("lid", CubeListBuilder.create().texOffs(0, 0).addBox(-6.08F, -12.74F, -5.14F, 12.0F, 9.0F, 12.0F, new CubeDeformation(0.0F)), PartPose.offset(0.0F, 0.78F, -0.78F));
+
+		PartDefinition head2 = shulker.addOrReplaceChild("head2", CubeListBuilder.create().texOffs(36, 0).addBox(-2.59F, 0.7F, -1.81F, 5.18F, 5.18F, 5.18F, new CubeDeformation(0.0F)), PartPose.offset(0.0F, -8.1F, -0.78F));
+
+		PartDefinition base = shulker.addOrReplaceChild("base", CubeListBuilder.create().texOffs(0, 21).addBox(-6.08F, -6.78F, -5.14F, 12.0F, 6.0F, 12.0F, new CubeDeformation(0.0F)), PartPose.offset(0.0F, 0.78F, -0.78F));
+
+		PartDefinition body = partdefinition.addOrReplaceChild("body", CubeListBuilder.create(), PartPose.offset(0.0F, 14.0F, 3.0F));
+
+		PartDefinition leftWing = body.addOrReplaceChild("leftWing", CubeListBuilder.create().texOffs(18, 39).addBox(0.0F, -6.5F, 0.0F, 9.0F, 13.0F, 0.0F, new CubeDeformation(0.0F)), PartPose.offset(8.0F, -0.5F, 0.0F));
+
+		PartDefinition leftWingTip = leftWing.addOrReplaceChild("leftWingTip", CubeListBuilder.create().texOffs(48, 10).addBox(0.0F, -6.5F, 0.0F, 8.0F, 13.0F, 0.0F, new CubeDeformation(0.0F)), PartPose.offset(9.0F, 0.0F, 0.0F));
+
+		PartDefinition rightWing = body.addOrReplaceChild("rightWing", CubeListBuilder.create().texOffs(0, 39).addBox(-9.0F, -6.5F, 0.0F, 9.0F, 13.0F, 0.0F, new CubeDeformation(0.0F)), PartPose.offset(-8.0F, -0.5F, 0.0F));
+
+		PartDefinition rightWingTip = rightWing.addOrReplaceChild("rightWingTip", CubeListBuilder.create().texOffs(36, 39).addBox(-8.0F, -6.5F, 0.0F, 8.0F, 13.0F, 0.0F, new CubeDeformation(0.0F)), PartPose.offset(-9.0F, 0.0F, 0.0F));
+
+		return LayerDefinition.create(meshdefinition, 64, 64);
+	}
+}