+++ Common/src/main/java/com/nyfaria/spookybats/entity/ShulkerBat.java
@@ -1,0 +1,19 @@
+package com.nyfaria.spookybats.entity;
+
+import com.nyfaria.spookybats.entity.api.SpookyBat;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.level.Level;
+
+public class ShulkerBat extends SpookyBat {
+	public ShulkerBat(EntityType<? extends SpookyBat> pEntityType, Level pLevel) {
+		super(pEntityType, pLevel);
+	}
+
+	public static AttributeSupplier.Builder createShulkerBatAttributes() {
+		return SpookyBat
+			.createBatAttributes()
+			.add(Attributes.MAX_HEALTH, 14);
+	}
+}